requirements.txt
fastapi
uvicorn
openai
python-dotenv
sqlalchemy
pydantic
pinecone-client
tree-sitter
tree-sitter-python
tree-sitter-javascript
asyncpg
alembic
python-jose
passlib
bcrypt
httpx
pytest
pytest-asyncio
